requirements.txt
cython>=0.29.28
python-dateutil>=2.8.1
dash>=2.0.0
dash_bootstrap_components>=1.0.0
pandas>=1.0.5
mlxtend>=0.19.0
gunicorn>=20.0.4
